Pulled chromium 77.0.3865.92

+++ chrome/browser/ui/bookmarks/bookmark_utils.h
@@ -88,7 +88,7 @@ int GetBookmarkDropOperation(Profile* profile,
                              const ui::DropTargetEvent& event,
                              const bookmarks::BookmarkNodeData& data,
                              const bookmarks::BookmarkNode* parent,
-                             int index);
+                             size_t index);
 
 // Returns true if the bookmark data can be dropped on |drop_parent| at
 // |index|. A drop from a separate profile is always allowed, where as
@@ -98,12 +98,11 @@ int GetBookmarkDropOperation(Profile* profile,
 bool IsValidBookmarkDropLocation(Profile* profile,
                                  const bookmarks::BookmarkNodeData& data,
                                  const bookmarks::BookmarkNode* drop_parent,
-                                 int index);
+                                 size_t index);
 
 #if defined(TOOLKIT_VIEWS)
 // |text_color| is the color of associated text and is used to derive the icon's
 // color.
-// TODO(estade): If Mac wants to use these, return gfx::Image instead.
 gfx::ImageSkia GetBookmarkFolderIcon(SkColor text_color);
 gfx::ImageSkia GetBookmarkManagedFolderIcon(SkColor text_color);
 #endif
