+++ components/webcrypto/algorithms/aes.cc
@@ -13,8 +13,8 @@
 #include "components/webcrypto/crypto_data.h"
 #include "components/webcrypto/jwk.h"
 #include "components/webcrypto/status.h"
-#include "third_party/WebKit/public/platform/WebCryptoAlgorithmParams.h"
-#include "third_party/WebKit/public/platform/WebCryptoKeyAlgorithm.h"
+#include "third_party/blink/public/platform/web_crypto_algorithm_params.h"
+#include "third_party/blink/public/platform/web_crypto_key_algorithm.h"
 
 namespace webcrypto {
 
@@ -63,7 +63,7 @@ Status AesAlgorithm::GenerateKey(const blink::WebCryptoAlgorithm& algorithm,
   if (status.IsError())
     return status;
 
-  unsigned int keylen_bits = algorithm.AesKeyGenParams()->LengthBits();
+  uint16_t keylen_bits = algorithm.AesKeyGenParams()->LengthBits();
 
   // 192-bit AES is intentionally unsupported (http://crbug.com/533699).
   if (keylen_bits == 192)
